feat(WorkList.tsx): migrate component
