+++ src/components/page/WorkListPage.tsx
@@ -1,11 +1,11 @@
 import React, { Fragment } from "react";
-import { Typography }     from "@material-ui/core";
 import { Add as AddIcon } from "@material-ui/icons";
 import gql                from "graphql-tag";
 import styled             from "styled-components";
 import { Query }          from "react-apollo";
 import arraysEqual              from "../../util/arraysEqual";
 import getTagsByURLQueryParam   from "../../util/getTagsByURLQueryParam";
+import { Work, WorkConnection } from "../../graphQL/type";
 import { PageComponentProps }   from "./../../App";
 import ErrorPage                from "../ErrorPage";
 import Fab                      from "../Fab";
@@ -14,7 +14,7 @@ import NotFound                 from "../NotFound";
 import Page                     from "../Page";
 import WorkDialog               from "../WorkDialog";
 import StreamSpinner            from "../StreamSpinner";
-import { Work, WorkConnection } from "../../graphQL/type";
+import WorkList                 from "../WorkList";
 
 interface State {
     paginationKey?: string;
@@ -151,27 +151,11 @@ export default class extends React.Component<PageComponentProps<{}>, State> {
 
                         return(
                             <Host>
-                                <WorkList>
-                                    {[...Array(this.state.workListRow).keys()].map(x => (
-                                        <div
-                                            key={x}
-                                            style={{
-                                                width: `calc(100% / ${this.state.workListRow})`
-                                            }}
-                                        >
-                                            {works
-                                                .filter((_, i) => i % this.state.workListRow === x)
-                                                .map(x => (
-                                                    <WorkItem
-                                                        work={x}
-                                                        key={x.id}
-                                                        onClick={this.handleClickOpen(x)}
-                                                    />
-                                                )
-                                            )}
-                                        </div>
-                                    ))}
-                                </WorkList>
+                                <WorkList
+                                    works={works}
+                                    workListRow={this.state.workListRow}
+                                    onWorkItemClick={this.handleClickOpen}
+                                />
                                 <WorkDialog
                                     history={history}
                                     open={this.state.workDialogVisible}
@@ -212,61 +196,4 @@ export default class extends React.Component<PageComponentProps<{}>, State> {
 const Host = styled.div`
     display: flex;
     flex-direction: column;
-`;
-
-const WorkList = styled.div`
-    margin: 0 3rem;
-    display: flex;
-    > * {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        margin: 0 1rem;
-    }
-`;
-
-const WorkItemBase = styled.div`
-    margin-bottom: 2rem;
-    width: 100%;
-`;
-
-// extends React.HTMLAttributes<HTMLDivElement>
-interface WorkItemProps extends React.HTMLAttributes<HTMLDivElement> {
-    work: Work;
-}
-
-const WorkItem = ({
-    work,
-    ...props
-}: WorkItemProps) => (
-    <WorkItemBase
-        {...props}
-    >
-        <WorkImage
-            src={(
-                work.imageUrl ? work.imageUrl
-              :                 "/img/no-image.png"
-            )}
-        />
-        <div>
-            <Typography variant="caption">
-                {work.user.displayName}
-            </Typography>
-            <Typography gutterBottom variant="title" component="h2">
-                {work.title}
-            </Typography>
-        </div>
-    </WorkItemBase>
-);
-
-const WorkImage = styled.img`
-    width: 100%;
-    border-radius: 8px;
-    transition: all 0.15s ease-in-out;
-    cursor: pointer;
-    :hover {
-        background-color: #fff;
-        transform: scale(1.1);
-        box-shadow: 0 7px 14px rgba(0,0,0,0.25), 0 5px 5px rgba(0,0,0,0.22);
-    }
 `;
